Adjusted module to implement more complex test
Adjusted module to expose I2S SCK directly on the `uo_out` output
pins and added a simple test to see if can read that
output correctly pulsing.

Note on `clk_divider`: As currently implemented, divides `fast_clk`
by a factor of `floor( ( factor + 2 ) / 2 ) * 2`. In this module's
default case, with a `factor` of 5, this leads to `i2s_sck` having
a period that is 6 `clk` periods. This means the calculation
written in the notes of `peripheral.v` that gets to a `factor` of 5
to ensure a minimum of 11.29 MHz actually leads to a `i2s_sck` that
has a frequency of 10.67 MHz.

As such, either `clk_divider` should subtract `factor` inputs by 2
automatically or there should be documentation to make it known that
the input to `factor` must be 2 less than the raw expected value to
ensure that the resultant `slow_clk` has at least the desired frequency
or higher.