Implemented rudimentary I2S transmitter
Implemented rudimentary, constantly-emitting I2S
transmitter on uo_out[1:3].

Took lots of debugging, mainly on trying
to transmit the MSB of the PISO shift register.

After much, much debugging, and a little desparate
help from AI, realized the following: firstly,
the original PISO code had a middleman register
that was unneeded (what AI contriubted).

More importantly, after realizing that the
gate-level testing vcd was more useful than it
originally seemed, found that the shift register
was taking longer to allow `x` to dissipate.
Essentially, after adding a longer delay before
actually testing, everything worked.

When trying to implementing the test, realized
that to poll on the i2s_sck, couldn't split
the uo_out signal to trigger upon a single
bit. Because of this, had to comment out
the tests in order to appease the Github
Actions gods.